Moved net_index_server from offst-index-server into offst-bin.
