+++ components/bin/src/stindex/mod.rs
@@ -1,3 +1,4 @@
+mod net_index;
 mod stindexlib;
 
 pub use self::stindexlib::{stindex, IndexServerBinError, StIndexCmd};
