+++ components/bin/src/stnode/stnodelib.rs
@@ -38,7 +38,7 @@ use proto::file::IdentityFile;
 use node::{NodeConfig, NodeState};
 
 use crate::stnode::file_trusted_apps::FileTrustedApps;
-use crate::stnode::node::{net_node, NetNodeError};
+use crate::stnode::net_node::{net_node, NetNodeError};
 
 /// Memory allocated to a channel in memory (Used to connect two components)
 const CHANNEL_LEN: usize = 0x20;
